pyproject.toml
[project]
name = "rag-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt>=0.9",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.33",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "rapidocr-onnxruntime>=1.4.4",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "unstructured[docx]>=0.18.14",
    "uvicorn[standard]>=0.35.0",
]
